% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar.R
\docType{data}
\name{polar}
\alias{polar}
\title{Polar bear litter size data set}
\format{
A data frame with 231 rows and 7 columns.
\describe{
  \item{year}{Catch year}
  \item{days}{Number of the day of the catch year}
  \item{id}{Unique specimen id}
  \item{age}{Age of the specimen, estimated using premolar tooth}
  \item{agecat}{Categorized age of the specimen}
  \item{length}{Body straight length (cm)}
  \item{cubnumber}{Litter size}
}
}
\source{
Folio, Dorinda Marie et al. (2019), Data from: How many cubs can a mum nurse?
Maternal age and size influence litter size in polar bears, Dryad, Dataset.
}
\usage{
data(polar)
}
\description{
This data set corresponds to live-captured polar bears from late March 1992 to beginning of May
2017 at Svalbard, Norway.
}
\examples{
data(polar)
head(polar)
}
\references{
Folio D. M., Aars J., Gimenez O., Derocher A. E., Wiig O. and Cubaynes S. (2019)
How many cubs can a mum nurse? Maternal age and size influence litter size in
polar bears, Biology letters, 15.
}
\keyword{datasets}
