% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/good-internal.R
\name{predict.glm.good}
\alias{predict.glm.good}
\alias{predict}
\title{Predict method for \code{glm.good} model fit}
\usage{
\method{predict}{glm.good}( object , newdata = NULL, se.fit = FALSE, ... )
}
\arguments{
\item{object}{object of class \code{glm.good}.}

\item{newdata}{An optional data frame in which to look for variables with which to predict.
If omitted, the fitted values are used.}

\item{se.fit}{logical, default FALSE. If TRUE, standard errors are provided.}

\item{...}{additional arguments.}
}
\value{
\code{predict.glm.good} produces a vector of predictions. If standard errors are required,
these are computed with the multivariate Delta method. If se.fit is TRUE, a list with the
components below is returned:
\item{fit}{vector with the fitted mean values}
\item{se.fit}{vector with standard errors of predicted mean values}
}
\description{
Predicted values for a \code{glm.good} object.
}
\keyword{internal}
