% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/good-internal.R
\name{summary.glm.good}
\alias{summary.glm.good}
\alias{summary}
\title{Summary of \code{glm.good} model fit}
\usage{
\method{summary}{glm.good}( object , ... )
}
\arguments{
\item{object}{object of class \code{glm.good}.}

\item{...}{additional arguments.}
}
\value{
\item{coefficients}{matrix of dimension p x 4 with columns given respectively the estimated coefficient,
standard errors, t-statistic and (two-sided) p-value.}
\item{loglik}{log-likelihood function of the fitted model.}
\item{aic}{Akaike Information Criterion value of the model.}
\item{bic}{Bayesian Information Criterion value of the model.}
\item{call}{unevaluated function call of the object}
\item{transformed}{point estimate and approximate standard error of \eqn{z=h^{-1}(z^*)} where \eqn{h^{-1}(\cdot)} is
the inverse of the link function and \eqn{z^{*}} is the corresponding estimated value given in coefficients.
The standard error is computed using the univariate Delta method. If the Good regression
has covariates, transformed returns a NULL.}
\item{testloglik}{log-likelihood function of the model without covariates (with intercept only).}
\item{LRT}{Likelihood Ratio Test statistic.}
\item{df}{degrees of freedom of the LRT.}
\item{pval}{p.value of the LRT.}
\item{testdist}{either Logarithmic or Geometric distributions tested against Good distribution with LRT.}
}
\description{
summary method for an object of class \code{glm.good}.
}
\keyword{internal}
